improved surface plot buttons, restructured 2obj
