+++ python/bertini_real/__init__.py
@@ -5,7 +5,6 @@ import bertini_real.curve
 import bertini_real.util
 import bertini_real.plot
 import bertini_real.glumpyplotter
-import bertini_real.tmesh
 import bertini_real.anaglypy
