+++ python/docs/conf.py
@@ -19,9 +19,9 @@ sys.path.insert(0, os.path.abspath('../bertini_real/'))
 
 # -- Project information -----------------------------------------------------
 
-project = u'Bertini_Real and Python'
-copyright = u'2019-2021, Dan Hessler, Foong Min Wong, Silviana Amethyst'
-author = u'Dan Hessler, Foong Min Wong, Silviana Amethyst'
+project = u'Bertini_real and Python'
+copyright = u'2019-2022, Silviana Amethyst'
+author = u'Silviana Amethyst, with Dan Hessler, Foong Min Wong, Caden Joergens'
 
 # The short X.Y version
 version = u''
