--- python/docs/tmesh.rst
@@ -1,13 +1,0 @@
-.. Test documentation master file, created by
-   sphinx-quickstart on Mon Feb 25 16:34:57 2019.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-bertinireal.tmesh
-================================
-
-Python libraries used: Trimesh, Numpy
-
-
-.. automodule:: tmesh
-   :members:
