+++ python/docs/tutorials/mesh_export.rst
@@ -1,10 +1,10 @@
-Exporting and solidifying stl files for 3d printing -- bertinireal.tmesh
+Exporting surfaces for 3d printing
 ===========================================================================
 
-The TMesh class in bertini_real allows user to export and solidify (raw & smooth surfaces) OBJ files for 3D printing.
+Bertini_real allows user to export and solidify (both raw & smooth surfaces) OBJ files for 3D printing, both for complete Surfaces, as well as Pieces of surfaces.
 
-Python Scripting
-*****************
+Exporting complete surfaces
+*****************************
 
 After decomposing a surface, you can run the following python codes to export and solidify OBJ in the Python Shell. 
 We are using a surface **"Whitney"** in this example ☂️ .
@@ -21,13 +21,13 @@ To export the raw version of OBJ, you can type:
 
 ::
 
-	bertini_real.tmesh.obj_raw(surface)
+	bertini_real.surface.export_obj_raw(surface)
 
 And you will see an output like this:
 
 ::
 
-	bertini_real.tmesh.obj_raw(surface)
+	bertini_real.surface.export_obj_raw(surface)
 
 	Generating raw OBJ surface...
 
@@ -35,72 +35,78 @@ And you will see an output like this:
 
 You can display OBJ using any 3D OBJ viewer:
 
-.. image:: tmesh_pictures/obj_raw_whitney.PNG
+.. image:: mesh_export_pictures/obj_raw_whitney.PNG
    :width: 300
 
-To export the smooth version of OBJ, you can type:
+
+
+To instead export the smooth version of OBJ, which requires that the surface has been sampled to your liking, you can type:
 
 ::
 
-	bertini_real.tmesh.obj_smooth(surface)
+	bertini_real.surface.obj_smooth(surface)
 
 And you will see an output like this:
 
 ::
 
-	bertini_real.tmesh.obj_smooth(surface)
+	bertini_real.surface.obj_smooth(surface)
 
 	Generating smooth OBJ surface...
 	Export obj_smooth_whitney.obj successfully
 
 You can display OBJ using any 3D OBJ viewer:
 
-.. image:: tmesh_pictures/obj_smooth_whitney.PNG
+.. image:: mesh_export_pictures/obj_smooth_whitney.PNG
    :width: 300
 
 To solidify the raw version of OBJ, you can type:
 
 ::
 
-	bertini_real.tmesh.solidify_raw(surface)
+	bertini_real.surface.solidify_raw(surface)
 
 And you will see an output like this:
 
 ::
 
-	bertini_real.tmesh.solidify_raw(surface)
+	bertini_real.surface.solidify_raw(surface)
 
 	Solidiying raw OBJ surface...
 	Export solidify_raw_whitney.obj successfully
 
 You can display OBJ using any 3D OBJ viewer:
 
-.. image:: tmesh_pictures/solidify_raw_whitney.PNG
+.. image:: mesh_export_pictures/solidify_raw_whitney.PNG
    :width: 300
 
 To solidify the smooth version of OBJ, you can type:
 
 ::
 
-	bertini_real.tmesh.solidify_smooth(surface)
+	bertini_real.surface.solidify_smooth(surface)
 
 And you will see an output like this:
 
 ::
 
-	bertini_real.tmesh.solidify_smooth(surface)
+	bertini_real.surface.solidify_smooth(surface)
 
 	Solidiying smooth OBJ surface...
 	Export solidify_smooth_whitney.obj successfully
 
 You can display OBJ using any 3D OBJ viewer:
 
-.. image:: tmesh_pictures/solidify_smooth_whitney.PNG
+.. image:: mesh_export_pictures/solidify_smooth_whitney.PNG
    :width: 300
 
-:Author:
-	Foong Min Wong
 
-:Version: 1.1 2019/09/17
+
+
+:Author:
+	Foong Min Wong, Silviana Amethyst
+
+:Version: 1.2 2022/03/05
 
 .. :Version: 1.0 2019/04/22
+
