+++ python/docs/tutorials/mesh_export_pictures/obj_raw_whitney.PNG
