+++ python/docs/tutorials/mesh_export_pictures/obj_smooth_whitney.PNG
