+++ python/docs/tutorials/mesh_export_pictures/solidify_raw_whitney.PNG
