+++ python/docs/tutorials/mesh_export_pictures/solidify_smooth_whitney.PNG
